Move csp to httputil
