+++ pkg/auth/handler/webapp/viewmodels/base.go
@@ -15,7 +15,6 @@ import (
 	"github.com/authgear/authgear-server/pkg/auth/webapp"
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/uiparam"
-	"github.com/authgear/authgear-server/pkg/lib/web"
 	"github.com/authgear/authgear-server/pkg/util/clock"
 	"github.com/authgear/authgear-server/pkg/util/geoip"
 	"github.com/authgear/authgear-server/pkg/util/httputil"
@@ -203,7 +202,7 @@ func (m *BaseViewModeler) ViewModel(r *http.Request, rw http.ResponseWriter) Bas
 		}
 	}
 
-	cspNonce := web.GetCSPNonce(r.Context())
+	cspNonce := httputil.GetCSPNonce(r.Context())
 
 	preferredLanguageTags := intl.GetPreferredLanguageTags(r.Context())
 	_, resolvedLanguageTagTag := intl.Resolve(preferredLanguageTags, string(m.DefaultLanguageTag), []string(m.SupportedLanguageTags))
