+++ pkg/lib/web/csp.go
@@ -1,147 +1,12 @@
 package web
 
 import (
-	"context"
 	"fmt"
 	"net/url"
 	"sort"
-	"strings"
+
+	"github.com/authgear/authgear-server/pkg/util/httputil"
 )
-
-type CSPSource interface {
-	CSPLevel() int
-	String() string
-}
-
-type CSPKeywordSourceLevel3 string
-
-var _ CSPSource = CSPKeywordSourceLevel3("")
-
-const (
-	CSPSourceStrictDynamic CSPKeywordSourceLevel3 = "'strict-dynamic'"
-)
-
-func (_ CSPKeywordSourceLevel3) CSPLevel() int {
-	return 3
-}
-
-func (s CSPKeywordSourceLevel3) String() string {
-	return string(s)
-}
-
-type CSPNonceSource struct {
-	Nonce string
-}
-
-var _ CSPSource = CSPNonceSource{}
-
-func (s CSPNonceSource) String() string {
-	return fmt.Sprintf("'nonce-%v'", s.Nonce)
-}
-
-func (s CSPNonceSource) CSPLevel() int {
-	return 2
-}
-
-type CSPSchemeSource struct {
-	Scheme string
-}
-
-var _ CSPSource = CSPSchemeSource{}
-
-func (s CSPSchemeSource) String() string {
-	return fmt.Sprintf("%v:", s.Scheme)
-}
-
-func (s CSPSchemeSource) CSPLevel() int {
-	return 1
-}
-
-type CSPHostSource struct {
-	Scheme string
-	Host   string
-}
-
-var _ CSPSource = CSPHostSource{}
-
-func (s CSPHostSource) String() string {
-	if s.Scheme != "" {
-		return fmt.Sprintf("%v://%v", s.Scheme, s.Host)
-	}
-	return fmt.Sprintf("%v", s.Host)
-}
-
-func (s CSPHostSource) CSPLevel() int {
-	return 1
-}
-
-type CSPKeywordSourceLevel1 string
-
-var _ CSPSource = CSPKeywordSourceLevel1("")
-
-const (
-	CSPSourceNone         CSPKeywordSourceLevel1 = "'none'"
-	CSPSourceSelf         CSPKeywordSourceLevel1 = "'self'"
-	CSPSourceUnsafeInline CSPKeywordSourceLevel1 = "'unsafe-inline'"
-)
-
-func (_ CSPKeywordSourceLevel1) CSPLevel() int {
-	return 1
-}
-
-func (s CSPKeywordSourceLevel1) String() string {
-	return string(s)
-}
-
-type CSPSources []CSPSource
-
-var _ sort.Interface = CSPSources{}
-
-func (s CSPSources) Len() int {
-	return len(s)
-}
-
-func (s CSPSources) Less(i, j int) bool {
-	// Higher level source must appear before lower level source.
-	return s[i].CSPLevel() > s[j].CSPLevel()
-}
-
-func (s CSPSources) Swap(i, j int) {
-	s[i], s[j] = s[j], s[i]
-}
-
-func (s CSPSources) String() string {
-	var strs []string
-	for _, source := range s {
-		strs = append(strs, source.String())
-	}
-	return strings.Join(strs, " ")
-}
-
-type dynamicCSPContextKeyType struct{}
-
-var dynamicCSPContextKey = dynamicCSPContextKeyType{}
-
-type cspNonceContextValue struct {
-	Nonce string
-}
-
-func WithCSPNonce(ctx context.Context, nonce string) context.Context {
-	v, ok := ctx.Value(dynamicCSPContextKey).(*cspNonceContextValue)
-	if ok {
-		v.Nonce = nonce
-		return ctx
-	}
-	return context.WithValue(ctx, dynamicCSPContextKey, &cspNonceContextValue{Nonce: nonce})
-}
-
-func GetCSPNonce(ctx context.Context) string {
-	v, ok := ctx.Value(dynamicCSPContextKey).(*cspNonceContextValue)
-	if ok {
-		return v.Nonce
-	}
-	return ""
-}
 
 type CSPDirectivesOptions struct {
 	PublicOrigin    string
@@ -154,35 +19,35 @@ type CSPDirectivesOptions struct {
 	AllowInlineScript bool
 }
 
-var wwwgoogletagmanagercom = CSPHostSource{
+var wwwgoogletagmanagercom = httputil.CSPHostSource{
 	Host: "www.googletagmanager.com",
 }
 
-var euassetsiposthogcom = CSPHostSource{
+var euassetsiposthogcom = httputil.CSPHostSource{
 	Host: "eu-assets.i.posthog.com",
 }
 
-var cdnjscloudflarecom = CSPHostSource{
+var cdnjscloudflarecom = httputil.CSPHostSource{
 	Host: "cdnjs.cloudflare.com",
 }
 
-var static2sharepointonlinecom = CSPHostSource{
+var static2sharepointonlinecom = httputil.CSPHostSource{
 	Host: "static2.sharepointonline.com",
 }
 
-var fontsgoogleapiscom = CSPHostSource{
+var fontsgoogleapiscom = httputil.CSPHostSource{
 	Host: "fonts.googleapis.com",
 }
 
-var fontsgstaticcom = CSPHostSource{
+var fontsgstaticcom = httputil.CSPHostSource{
 	Host: "fonts.gstatic.com",
 }
 
-var challengescloudflarecom = CSPHostSource{
+var challengescloudflarecom = httputil.CSPHostSource{
 	Host: "challenges.cloudflare.com",
 }
 
-var wwwgooglecom = CSPHostSource{
+var wwwgooglecom = httputil.CSPHostSource{
 	Host: "www.google.com",
 }
 
@@ -192,24 +57,24 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 		return nil, err
 	}
 
-	baseSrc := CSPSources{CSPSourceSelf}
+	baseSrc := httputil.CSPSources{httputil.CSPSourceSelf}
 	if opts.CDNHost != "" {
-		baseSrc = append(baseSrc, CSPHostSource{
+		baseSrc = append(baseSrc, httputil.CSPHostSource{
 			Host: opts.CDNHost,
 		})
 	}
 
 	// Unsafe-inline gets ignored if nonce is provided
 	// https://w3c.github.io/webappsec-csp/#allow-all-inline
-	var scriptSrc CSPSources
+	var scriptSrc httputil.CSPSources
 	if opts.AllowInlineScript {
-		scriptSrc = CSPSources{
-			CSPSourceUnsafeInline,
+		scriptSrc = httputil.CSPSources{
+			httputil.CSPSourceUnsafeInline,
 		}
 	} else {
-		scriptSrc = CSPSources{
-			CSPSourceStrictDynamic,
-			CSPNonceSource{
+		scriptSrc = httputil.CSPSources{
+			httputil.CSPSourceStrictDynamic,
+			httputil.CSPNonceSource{
 				Nonce: opts.Nonce,
 			},
 		}
@@ -220,7 +85,7 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 		euassetsiposthogcom,
 		challengescloudflarecom,
 		wwwgooglecom,
-		CSPHostSource{
+		httputil.CSPHostSource{
 			Scheme: "https",
 			Host:   "browser.sentry-cdn.com",
 		},
@@ -228,14 +93,14 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 	scriptSrc = append(scriptSrc, baseSrc...)
 	sort.Sort(scriptSrc)
 
-	frameSrc := CSPSources{
+	frameSrc := httputil.CSPSources{
 		wwwgoogletagmanagercom,
 		challengescloudflarecom,
 		wwwgooglecom,
-		CSPSourceSelf,
+		httputil.CSPSourceSelf,
 	}
 
-	fontSrc := CSPSources{
+	fontSrc := httputil.CSPSources{
 		cdnjscloudflarecom,
 		static2sharepointonlinecom,
 		fontsgoogleapiscom,
@@ -244,8 +109,8 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 	fontSrc = append(fontSrc, baseSrc...)
 	sort.Sort(fontSrc)
 
-	styleSrc := CSPSources{
-		CSPSourceUnsafeInline,
+	styleSrc := httputil.CSPSources{
+		httputil.CSPSourceUnsafeInline,
 		cdnjscloudflarecom,
 		wwwgoogletagmanagercom,
 		fontsgoogleapiscom,
@@ -253,16 +118,16 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 	styleSrc = append(styleSrc, baseSrc...)
 	sort.Sort(styleSrc)
 
-	imgSrc := CSPSources{
-		CSPSchemeSource{
+	imgSrc := httputil.CSPSources{
+		httputil.CSPSchemeSource{
 			Scheme: "http",
 		},
-		CSPSchemeSource{
+		httputil.CSPSchemeSource{
 			Scheme: "https",
 		},
 		// We use data URI to show QR image.
 		// We can display external profile picture.
-		CSPSchemeSource{
+		httputil.CSPSchemeSource{
 			Scheme: "data",
 		},
 	}
@@ -271,17 +136,17 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 
 	// 'self' does not include websocket in Safari :(
 	// https://github.com/w3c/webappsec-csp/issues/7
-	connectSrc := CSPSources{
-		CSPSourceSelf,
-		CSPHostSource{
+	connectSrc := httputil.CSPSources{
+		httputil.CSPSourceSelf,
+		httputil.CSPHostSource{
 			Scheme: "https",
 			Host:   "www.google-analytics.com",
 		},
-		CSPHostSource{
+		httputil.CSPHostSource{
 			Scheme: "ws",
 			Host:   u.Host,
 		},
-		CSPHostSource{
+		httputil.CSPHostSource{
 			Scheme: "wss",
 			Host:   u.Host,
 		},
@@ -296,22 +161,22 @@ func CSPDirectives(opts CSPDirectivesOptions) ([]string, error) {
 		sentryDSNHost = u.Host
 	}
 	if sentryDSNHost != "" {
-		connectSrc = append(connectSrc, CSPHostSource{
+		connectSrc = append(connectSrc, httputil.CSPHostSource{
 			Host: sentryDSNHost,
 		})
 	}
 	sort.Sort(connectSrc)
 
-	var frameAncestors CSPSources
+	var frameAncestors httputil.CSPSources
 	if len(opts.FrameAncestors) > 0 {
 		for _, host := range opts.FrameAncestors {
-			frameAncestors = append(frameAncestors, CSPHostSource{
+			frameAncestors = append(frameAncestors, httputil.CSPHostSource{
 				Host: host,
 			})
 		}
 	} else {
-		frameAncestors = CSPSources{
-			CSPSourceNone,
+		frameAncestors = httputil.CSPSources{
+			httputil.CSPSourceNone,
 		}
 	}
